Fix unifying static and dynamic variables
Add restriction for both static and dynamic variables to only track
variables with a certain size (the size if hard-coded temporarily).

Fix problems with splay tree. Use pointer to root instead of root.

Add more debugging info

+++ src/tool/hpcfnbounds/variable-entries.cpp
@@ -184,8 +184,8 @@ dump_variable_entry(void *addr, long size, const char *comment)
   num_entries_total += 2;
 
   if (server_mode()) {
-    syserv_add_addr(addr, 2*variable_entries.size());
-    syserv_add_addr((void *)size, 2*variable_entries.size());
+    syserv_add_addr(addr, 2 * (long)variable_entries.size());
+    syserv_add_addr((void *)size, 2 * (long)variable_entries.size());
     return;
   }
